manager/requirements.txt
discord.py>=1.1.1
aiohttp>=3.6.0,<3.7.0
websockets>=6.0,!=7.0,!=8.0,!=8.0.1,<9.0
google-cloud-error-reporting>=0.33.0
google-cloud-firestore>=1.9.0
pytz>=2019.2